Move websocket server code in own module
