+++ server/glsp-server-websocket/pom.xml
@@ -7,9 +7,9 @@
 		<artifactId>glsp-parent</artifactId>
 		<version>1.2.0-SNAPSHOT</version>
 	</parent>
-	<artifactId>glsp-server</artifactId>
-	<name>glsp-server</name>
-	<description>Generic server component for GLSP</description>
+	<artifactId>glsp-server-websocket</artifactId>
+	<name>glsp-server-websocket</name>
+	<description>Websocket-based GLSP Server implementation</description>
 	<url>https://github.com/eclipsesource/graphical-lsp</url>
 	<licenses>
 		<license>
@@ -60,7 +60,7 @@
 	<dependencies>
 		<dependency>
 			<groupId>com.eclipsesource.glsp</groupId>
-			<artifactId>glsp-api</artifactId>
+			<artifactId>glsp-server</artifactId>
 			<version>1.2.0-SNAPSHOT</version>
 		</dependency>
 		<dependency>
@@ -73,5 +73,11 @@
 			<artifactId>org.eclipse.sprotty</artifactId>
 			<version>0.7.0-SNAPSHOT</version>
 		</dependency>
+		<dependency>
+			<groupId>org.eclipse.jetty.websocket</groupId>
+			<artifactId>javax-websocket-server-impl</artifactId>
+			<version>9.4.10.v20180503</version>
+		</dependency>
+
 	</dependencies>
 </project>
