+++ server/glsp-server-websocket/src/main/java/com/eclipsesource/glsp/server/websocket/LoggingServerEndpoint.java
